fetch master branch from forked repo

+++ java/java-tests/testData/codeInsight/daemonCodeAnalyzer/quickFix/castMethodParameters15/afterVarargsLessArgs.java
@@ -2,7 +2,7 @@
 class a {
     void f(Long l, String... s) {}
     void g() {
-        f((long) 0);
+        f(0L);
     }
 }
